Replace files content
